@@ -1,43 +1,54 @@
 {
 	"name": "friendconstellation",
-	"version": "0.0.1",
 	"private": true,
+	"version": "2.0.0",
+	"type": "module",
 	"scripts": {
 		"dev": "vite dev",
 		"build": "vite build",
 		"preview": "vite preview",
+		"db:dev": "npx prisma migrate dev",
+		"db:deploy": "npx prisma migrate deploy",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+		"format": "prettier --write .",
 		"lint": "prettier --check . && eslint .",
-		"format": "prettier --write ."
+		"test:unit": "vitest",
+		"test": "npm run test:unit -- --run"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.0.0",
+		"@eslint/compat": "^1.2.3",
+		"@sveltejs/adapter-node": "^5.2.9",
 		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"@types/eslint": "^8.56.0",
-		"@typescript-eslint/eslint-plugin": "^7.0.0",
-		"@typescript-eslint/parser": "^7.0.0",
-		"carbon-icons-svelte": "^12.6.0",
-		"carbon-preprocess-svelte": "^0.11.0",
-		"eslint": "^8.56.0",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
+		"@types/node": "^22.10.2",
+		"autoprefixer": "^10.4.20",
+		"bits-ui": "^0.22.0",
+		"clsx": "^2.1.1",
+		"eslint": "^9.7.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
-		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
-		"prisma": "^5.11.0",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.6.0",
-		"tslib": "^2.4.1",
+		"eslint-plugin-svelte": "^2.36.0",
+		"globals": "^15.0.0",
+		"prettier": "^3.3.2",
+		"prettier-plugin-svelte": "^3.2.6",
+		"prettier-plugin-tailwindcss": "^0.6.5",
+		"prisma": "^6.1.0",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
+		"tailwind-merge": "^2.6.0",
+		"tailwind-variants": "^0.3.0",
+		"tailwindcss": "^3.4.9",
 		"typescript": "^5.0.0",
-		"vite": "^5.0.3"
+		"typescript-eslint": "^8.0.0",
+		"vite": "^5.4.11",
+		"vitest": "^2.0.4"
 	},
-	"type": "module",
 	"dependencies": {
-		"@prisma/client": "^5.11.0",
-		"@sveltejs/adapter-node": "^5.0.1",
-		"axios": "^1.6.8",
-		"carbon-components-svelte": "^0.85.0",
-		"dotenv": "^16.4.5"
+		"@prisma/client": "6.1.0",
+		"@tailwindcss/forms": "^0.5.9",
+		"@tailwindcss/typography": "^0.5.15",
+		"dotenv": "^16.4.7",
+		"lucide-svelte": "^0.469.0",
+		"uuid": "^11.0.3"
 	}
 }
